+++ NEWS.md
@@ -1,3 +1,4 @@
+- Added `expand = TRUE` argument to `ypr_populations()` to allow control over all combinations.
 - Added parameters `k2`, `Linf2` and `L2` to allow for biphasic growth.
 - Replaced parameters 
     - `M` (annual instantaneous natural mortality rate)